requirements.txt
aiohttp
lxml
pyyaml
